WIP OpenWebUi With PgSQL

+++ generate_compose.sh
@@ -13,6 +13,7 @@ if ! command -v jq &>/dev/null; then
           exit 1
       fi
   elif [[ "$OSTYPE" == "cygwin" || "$OSTYPE" == "msys" || "$OSTYPE" == "win32" ]]; then
+    if [[ ! -f "jq.exe" ]]; then
       # Installation de jq sur Windows
       echo "Téléchargement de jq pour Windows..."
       curl -L -o jq.exe https://github.com/jqlang/jq/releases/latest/download/jq-win64.exe
@@ -23,43 +24,55 @@ if ! command -v jq &>/dev/null; then
           exit 1
       fi
 
-      echo "Déplacement de jq dans le répertoire cible..."
-      mv jq.exe "/c/Program Files/Git/usr/bin/jq.exe"
-
       # Vérifier si le déplacement a réussi
-      if [[ ! -f "/c/Program Files/Git/usr/bin/jq.exe" ]]; then
+      if [[ ! -f "./jq.exe" ]]; then
           echo "Erreur lors du déplacement de jq.exe."
           exit 1
       fi
 
       echo "jq a été installé avec succès."
+    fi
   else
       echo "Système d'exploitation non pris en charge pour l'installation automatique de jq."
       exit 1
   fi
 fi
 
+#!/bin/bash
+
 # Lire le fichier config.json
-config_json=$(cat config.json)
+configJson=$(cat config.json)
 
 # Extraire les variables du JSON
-INSTANCES=$(echo $config_json | jq -r '.INSTANCES[]')
-GPUS_COUNT=$(echo $config_json | jq -r '.GPUS_COUNT')
-SHARED_PATH_HOST=$(echo $config_json | jq -r '.SHARED_PATH_HOST')
-CUSTOM_API_KEY=$(echo $config_json | jq -r '.CUSTOM_API_KEY')
-POSTGRES_USER=$(echo $config_json | jq -r '.POSTGRES_USER')
-POSTGRES_PASSWORD=$(echo $config_json | jq -r '.POSTGRES_PASSWORD')
-POSTGRES_DB=$(echo $config_json | jq -r '.POSTGRES_DB')
+
+INSTANCES=$(./jq.exe -r '.INSTANCES[]' "config.json")
+GPUS_COUNT=$(./jq.exe -r '.GPUS_COUNT' "config.json")
+SHARED_PATH_HOST=$(./jq.exe -r '.SHARED_PATH_HOST' "config.json")
+CUSTOM_API_KEY=$(./jq.exe -r '.CUSTOM_API_KEY' "config.json")
+POSTGRES_USER=$(./jq.exe -r '.POSTGRES_USER' "config.json")
+POSTGRES_PASSWORD=$(./jq.exe -r '.POSTGRES_PASSWORD' "config.json")
+POSTGRES_DB=$(./jq.exe -r '.POSTGRES_DB' "config.json")
+RESTART=$(./jq.exe -r '.RESTART_OPTION' "config.json")
+WEBUI_DOCKER_TAG=$(./jq.exe -r '.WEBUI_DOCKER_TAG' "config.json")
+WEBUI_PORT=$(./jq.exe -r '.WEBUI_PORT' "config.json")
+
+if [[ -z "$WEBUI_DOCKER_TAG" ]]; then
+    WEBUI_DOCKER_TAG="main"
+fi
+if [[ -z "$WEBUI_PORT" ]]; then
+    WEBUI_PORT="3000"
+fi
 
 # Vérifier si l'image Docker 'crewai-studio' existe localement
-if ! docker images | grep -q 'crewai-studio'; then
+imageExists=$(docker images -q crewai-studio)
+if [ -z "$imageExists" ]; then
     echo "L'image 'crewai-studio' n'existe pas localement. Clonage du dépôt et construction de l'image..."
 
     # Cloner le dépôt GitHub
     git clone https://github.com/strnad/CrewAI-Studio.git
 
     # Construire l'image Docker avec le nom 'crewai-studio'
-    cd CrewAI-Studio || exit
+    cd CrewAI-Studio
     docker build -t crewai-studio .
 
     # Supprimer le dossier CrewAI-Studio après la construction
@@ -70,24 +83,24 @@ else
 fi
 
 # Initialiser le contenu du fichier docker-compose.yaml
-compose_content="version: '3.8'\n\nservices:\n"
-
-# Ajouter le service Ollama au docker-compose
-compose_content+="
+composeContent=$(cat <<EOF
+services:
   ollama:
     image: ollama/ollama:latest
-    environment:
-      - OLLAMA_API_KEY=${CUSTOM_API_KEY}
+    restart: $RESTART
     ports:
-      - \"11434:11434\"
+      - "11434:11434"
     volumes:
       - ollama:/root/.ollama
-      - ${SHARED_PATH_HOST}:/shared
-"
+      - $SHARED_PATH_HOST:/shared
+EOF
+)
 
 # Ajouter la configuration GPU si GPUS_COUNT est positif ou "all"
-if [[ "$GPUS_COUNT" -gt 0 || "$GPUS_COUNT" == "all" ]]; then
-    compose_content+="
+if [[ "$GPUS_COUNT" -gt 0 ]]; then
+    composeContent+=$(cat <<EOF
+
+    environment:
       - NVIDIA_VISIBLE_DEVICES=all
       - NVIDIA_DRIVER_CAPABILITIES=compute,utility
       - CUDA_VISIBLE_DEVICES=0
@@ -95,76 +108,102 @@ if [[ "$GPUS_COUNT" -gt 0 || "$GPUS_COUNT" == "all" ]]; then
       resources:
         reservations:
           devices:
-            - driver: nvidia"
-
-    if [[ "$GPUS_COUNT" == "all" ]]; then
-        compose_content+="\n              count: all"
-    else
-        compose_content+="\n              count: $GPUS_COUNT"
-    fi
-
-    compose_content+="\n              capabilities: [gpu]"
+            - driver: nvidia
+              count: $GPUS_COUNT
+              capabilities: [gpu]
+EOF
+)
 fi
 
-# Transformer la liste des instances pour le format souhaité
-OLLAMA_MODELS=$(echo $INSTANCES | tr ' ' '\n' | awk '{print "ollama/"$0}' | paste -sd, -)
+$ln=$( IFS=$','; echo "${INSTANCES[*]}" )
 
-# Ajouter le service crewai-studio au docker-compose
-compose_content+="\n  crewai-studio:\n"
-compose_content+="    image: crewai-studio\n"
-compose_content+="    ports:\n"
-compose_content+="      - \"8501:8501\"\n"
-compose_content+="    volumes:\n"
-compose_content+="      - crewai-data:/data\n"
-compose_content+="    environment:\n"
-compose_content+="      - DB_URL=postgresql://${POSTGRES_USER}:${POSTGRES_PASSWORD}@db:5432/${POSTGRES_DB}\n"
-compose_content+="      - OLLAMA_HOST=http://ollama:11434\n"
-compose_content+="      - OLLAMA_MODELS=${OLLAMA_MODELS}\n"
-compose_content+="    depends_on:\n"
-compose_content+="      - ollama\n"
+echo  $ln;
+ln=${ln//[$'\t\r\n ']};
+echo  $ln;
 
-# Ajouter le service PostgreSQL au docker-compose
-compose_content+="\n  db:\n"
-compose_content+="    image: postgres:latest\n"
-compose_content+="    environment:\n"
-compose_content+="      POSTGRES_USER: ${POSTGRES_USER}\n"
-compose_content+="      POSTGRES_PASSWORD: ${POSTGRES_PASSWORD}\n"
-compose_content+="      POSTGRES_DB: ${POSTGRES_DB}\n"
-compose_content+="    ports:\n"
-compose_content+="      - \"5432:5432\"\n"
-compose_content+="    volumes:\n"
-compose_content+="      - postgres-data:/var/lib/postgresql/data\n"
+composeContent+=$(cat <<EOF
 
-compose_content+="\nvolumes:\n  ollama:\n  crewai-data:\n  postgres-data:\n"
+  postgres-db:
+    image: postgres:latest
+    restart: ${RESTART}
+    environment:
+      POSTGRES_USER: ${POSTGRES_USER}
+      POSTGRES_PASSWORD: ${POSTGRES_PASSWORD}
+      POSTGRES_DB: ${POSTGRES_DB}
+    ports:
+      - "5432:5432"
+    volumes:
+      - postgres-data:/var/lib/postgresql/data
+  crewai-studio:
+    image: crewai-studio
+    restart: ${RESTART}
+    ports:
+      - "8501:8501"
+    volumes:
+      - crewai-data:/data
+    environment:
+      - DB_URL=postgresql://${POSTGRES_USER}:${POSTGRES_PASSWORD}@postgres-db:5432/${POSTGRES_DB}
+      - OLLAMA_HOST=http://ollama:11434
+      - OLLAMA_MODELS=$ln
+    depends_on:
+      - ollama
+      - postgres-db
+EOF
+)
+
+# Demander confirmation pour ajouter une interface pour accéder à l'IA
+read -p "Voulez-vous ajouter une interface pour accéder à l'IA ? (y/n*)" confirmationUI
+if [ "$confirmationUI" == "y" ]; then
+    composeContent+=$(cat <<EOF
+
+  open-webui:
+    build:
+      context: .
+      args:
+        OLLAMA_BASE_URL: '/ollama'
+      dockerfile: Dockerfile
+    image: ghcr.io/open-webui/open-webui:${WEBUI_DOCKER_TAG}
+    container_name: open-webui
+    volumes:
+      - open-webui:/app/backend/data
+    depends_on:
+      - ollama
+    ports:
+      - ${WEBUI_PORT}:8080
+    environment:
+      - 'OLLAMA_BASE_URL=http://ollama:11434'
+      - 'WEBUI_SECRET_KEY='
+    extra_hosts:
+      - host.docker.internal:host-gateway
+    restart: ${RESTART}
+EOF
+)
+fi
+
+composeContent+=$(cat <<EOF
+
+volumes:
+  ollama:
+  crewai-data:
+  postgres-data:
+EOF
+)
+
+if [ "$confirmationUI" == "y" ]; then
+    composeContent+=$(cat <<EOF
+
+  open-webui:
+EOF
+)
+fi
 
 # Écrire le contenu dans le fichier docker-compose.yaml
-echo -e "$compose_content" > docker-compose.yaml
+echo "$composeContent" > docker-compose.yaml
 
 echo "Fichier docker-compose.yaml généré avec succès !"
 
-# Demander confirmation pour lancer les conteneurs
-read -p "Voulez-vous lancer les conteneurs Docker et télécharger les modèles ? (y/n) " confirmation
-if [[ "$confirmation" == "y" ]]; then
-    # Lancer docker-compose en arrière-plan
-    docker-compose up -d
-
-    # Attendre 10 secondes pour s'assurer que tous les conteneurs sont démarrés
-    echo "Attente de 10 secondes pour s'assurer que tous les conteneurs sont démarrés..."
-    sleep 10
-
-    # Envoyer une requête HTTP POST pour chaque modèle
-    port=11434
-    queryUrl="http://localhost:$port/api/pull"
-
-    for instance in $INSTANCES; do
-        instance=$(echo $instance | tr -d '\r')
-        queryBody=$(jq -n --arg model "$instance" '{"model": $model}')
-
-        echo "Téléchargement du modèle: $instance"
-
-        # Envoyer la requête POST
-        curl -s -o /dev/null -X POST $queryUrl -H "Content-Type: application/json" -d "$queryBody"
-    done
-
-    echo "Initialisation des conteneurs terminée."
+# Demander confirmation pour lancer les conteneurs Docker et télécharger les modèles
+read -p "Voulez-vous lancer les conteneurs Docker et télécharger les modèles ? (y/n*) " confirmation
+if [ "$confirmation" == "y" ]; then
+    ./start_compose.sh
 fi
